backend/requirements.txt
# Flask and extensions
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0

# Database
psycopg2-binary==2.9.9

# OpenAI
openai>=1.12.0

# Qdrant Vector DB
qdrant-client==1.7.0

# Utilities
python-dotenv==1.0.0
marshmallow==3.20.1
qrcode[pil]==7.4.2
Pillow==10.1.0
uuid==1.30

# Security
bcrypt==4.1.2
PyJWT==2.8.0

# Development
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.1
flake8==6.1.0

